Use 2-cores workers for make_examples step
Due to high preemption rate of single core machines, specially when the
zone we use to run the job is more busy, I decided to use 2-cores
workers. This might increase the cost a bit (comparing to finishing job
using single core machines) however we avoid cases the overal cost of
preemptive machines approaches the cost of non-preemptive machines.